Alpha feedback
I think there needs to be a fair amount of work done to get a level editor going, and I would suggest coding duties for it are not put on one person. I think a lot of work needs to go into the UI for this as well, as I have yet to see a game over screen or any main menu for the game. The character models are fantastic and I think the game has a lot of potential, but I could not play anything but the sandbox due to some merge conflicts which were unfortunately pushed to the main repository. 
Alpha score: 85/100